Relax an overzealous safety check on pointer arithmetic.
Previously, the check would only succeed if the result of the addition
was an in bounds pointer, or if the result was NULL.  However, it is
also valid to, e.g., add offsets to NULL (which is just the 0
bitpattern).  Here, we relax the restriction so that the result must
be a valid pointer OR any ordinary bitvector value.

This allows certain patterns that clang produces via loop-rearranging
where a GEP might be applied to the NULL pointer, even though the
source program appears like it would skip doing the arithmetic.